fixing user and pet seeds
